Use defaults set in option structs.
Previously, defaults in the options structs were not used and had to be
set in the init functions. This was due to a bug in go-flags that has
since been fixed in
jessevdk/go-flags@8e054df750e8371c394e95da0a852e29212dba45

These now are set in the option structs for both gets and puts.

The defaults for the path are set in code to os.Stdin and os.Stdout
rather than using /dev/stdin and /dev/stdout in the option structs since that is not
supported cross-platform.