Fix request signature for non-urlsafe object names
Object names that would be transformed by url percent encoding would fail with the error

```
"The request signature we calculated does not match the signature you provided. Check your key and signing method."
```

In these cases, the canonicalized URI used for signing is different from the URI the signature is sent to.  The initial quoting is done here:

https://github.com/rlmcpherson/s3gof3r/blob/c8635b85d6b6d05ff0a959fd5b71c53305e72c11/putter.go#L81