Use AWS_REGION environment variable if it is set
Use AWS_REGION environment variable if it is set.

Without this change, trying to put to a bucket in a region other than
us-east-1 results in something like:

> gof3r error: 400: "The authorization header is malformed; the region
> 'us-east-1' is wrong; expecting 'us-west-2'"

I only have buckets in us-west-2, and I have verified that this
change (and export AWS_REGION=us-west-2) allows me to `put`, `cp`, and
`get` objects from those buckets.

I am not sure how it interacts with some of the other features.